New Wavehandler script
Wavehandler has a spawnerprefab that it will spawn at random set location. It also has an difficulty that will change the behaviour of the enemies.
A wavenumber to indicate which wave we are currently on, and a TimeBetweenWaves that is the time between each wave. Before that starts ticking the OnWaveFinished will be invoked and can update the UI and stuff. When is has reached 0 the new wave will begin and spawner object are instantiated and the OnNewWave will be invoked to update UI and stuff.